+++ app/helpers/dorsale/comments_helper.rb
@@ -1,10 +1,8 @@
 module Dorsale::CommentsHelper
-  def comments_for(commentable, options = {})
-    comments = policy_scope(::Dorsale::Comment)
-      .where(commentable: commentable)
-      .preload(:commentable, :author)
-
-    new_comment = comments.new(author: current_user)
+  def comments_for(commentable, comments = nil)
+    comments    = ::Dorsale::Comment.where(commentable: commentable) if comments.nil?
+    comments    = policy_scope(comments).preload(:commentable, :author)
+    new_comment = policy_scope(comments).new(commentable: commentable, author: current_user)
 
     render(
       :partial => "dorsale/comments/comments",