+++ spec/models/dorsale/customer_vault/person_spec.rb
@@ -24,7 +24,16 @@ RSpec.describe ::Dorsale::CustomerVault::Person, type: :model do
         expect { link.reload }.to raise_error(ActiveRecord::RecordNotFound)
       end
     end
-
   end # describe '#links'
+
+  it "should return self_and_related_comments" do
+    corporation         = create(:customer_vault_corporation)
+    individual          = create(:customer_vault_individual, corporation: corporation)
+    corporation_comment = create(:dorsale_comment, commentable: corporation)
+    individual_comment  = create(:dorsale_comment, commentable: individual)
+
+    expect(corporation.self_and_related_comments).to contain_exactly(corporation_comment, individual_comment)
+    expect(individual.self_and_related_comments).to contain_exactly(individual_comment)
+  end
 
 end