+++ app/helpers/dorsale/all_helpers.rb
@@ -7,6 +7,7 @@ module Dorsale::AllHelpers
   include ::Dorsale::CommentsHelper
   include ::Dorsale::Alexandrie::AttachmentsHelper
   include ::Dorsale::BillingMachine::ApplicationHelper
+  include ::Dorsale::CustomerVault::ApplicationHelper
   include ::Dorsale::Flyboy::ApplicationHelper
   include ::Dorsale::ExpenseGun::ApplicationHelper
   include ::Dorsale::UsersHelper