+++ app/helpers/dorsale/customer_vault/application_helper.rb
@@ -13,4 +13,30 @@ module Dorsale::CustomerVault::ApplicationHelper
       [Dorsale::CustomerVault::Individual.t,  "Dorsale::CustomerVault::Individual"],
     ]
   end
+
+  def person_address_blank?(person)
+    [
+      person.address.street,
+      person.address.street_bis,
+      person.address.zip,
+      person.address.city,
+      person.address.country,
+    ].all?(&:blank?)
+  end
+
+  def person_social_blank?(person)
+    [
+      person.skype,
+      person.www,
+      person.twitter,
+      person.facebook,
+      person.linkedin,
+      person.viadeo,
+      person.try(:societe_com),
+    ].all?(&:blank?)
+  end
+
+  def person_related_people_blank?(person)
+    person.individuals.empty?
+  end
 end